+++ src/main/java/com/adobe/campaign/tests/service/CallContent.java
@@ -6,6 +6,7 @@ import com.adobe.campaign.tests.service.exceptions.TargetJavaMethodCallException
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.Arrays;
@@ -74,6 +75,21 @@ public class CallContent {
         return lr_method.get(0);
     }
 
+    /**
+     * Returns the method object that is defined by this class
+     *
+     * @return the method object
+     */
+    public Constructor fetchConstructor(Class in_class) {
+
+        List<Constructor> lr_method = fetchConstructorCandidates(in_class);
+        if (lr_method.size() > 1) {
+            throw new AmbiguousMethodException(
+                    "We could not find a unique method for " + this.getFullName());
+        }
+        return lr_method.get(0);
+    }
+
     public List<Method> fetchMethodCandidates(Class in_class) {
 
         List<Method> lr_method = lr_method = Arrays.stream(in_class.getMethods())
@@ -84,6 +100,21 @@ public class CallContent {
         if (lr_method.size() == 0) {
             throw new NonExistentJavaObjectException(
                     "Method " + this.getClassName() + "." + this.getMethodName() + "   with " + this.getArgs().length
+                            + " arguments could not be found.");
+        }
+        return lr_method;
+    }
+
+    public List<Constructor> fetchConstructorCandidates(Class in_class) {
+
+        List<Constructor> lr_method = Arrays.stream(in_class.getConstructors())
+                .filter(fp -> fp.getParameterCount() == this.getArgs().length).collect(
+                        Collectors.toList());
+
+        if (lr_method.size() == 0) {
+            throw new NonExistentJavaObjectException(
+                    "Constructor " + this.getClassName() + "." + this.getMethodName() + "   with "
+                            + this.getArgs().length
                             + " arguments could not be found.");
         }
         return lr_method;
@@ -111,12 +142,24 @@ public class CallContent {
                                 .substring(0, this.getClassName().lastIndexOf('.')) : this.getClassName());
             }
 
-            Class ourClass = Class.forName(getClassName(), true, iClassLoader);
+            Object l_instanceObject = iClassLoader.getCallResultCache().get(this.getClassName());
 
-            Method l_method = fetchMethod(ourClass);
+            String l_usedClassName = (l_instanceObject == null ? this.getClassName() :
+                    iClassLoader.getCallResultCache()
+                            .get(this.getClassName()).getClass().getTypeName());
 
-            Object ourInstance = ourClass.getDeclaredConstructor().newInstance();
-            lr_object = l_method.invoke(ourInstance, expandArgs(iClassLoader));
+            Class ourClass = Class.forName(l_usedClassName, true, iClassLoader);
+
+            if (isConstructorCall()) {
+                Constructor l_constructor = fetchConstructor(ourClass);
+                lr_object = l_constructor.newInstance(expandArgs(iClassLoader));
+            } else {
+                Method l_method = fetchMethod(ourClass);
+
+                Object ourInstance = (l_instanceObject == null) ? ourClass.getDeclaredConstructor().newInstance() : l_instanceObject;
+                lr_object = l_method.invoke(ourInstance, expandArgs(iClassLoader));
+            }
+
         } catch (IllegalArgumentException e) {
             throw new NonExistentJavaObjectException(
                     "The given method " + this.getFullName() + " could not accept the given arguments..");
@@ -175,5 +218,16 @@ public class CallContent {
     public Object[] expandArgs(IntegroBridgeClassLoader in_currentClassLoader) {
         return Arrays.stream(getArgs()).map(arg -> in_currentClassLoader.getCallResultCache().getOrDefault(
                 arg, arg)).toArray();
+    }
+
+    /**
+     * Lets us know if the giben Call Content is a Constructor call
+     *
+     * @return true if the methos is actually a constructor
+     */
+    @JsonIgnore
+    public boolean isConstructorCall() {
+        return getMethodName() == null || getClassName().substring(getClassName().lastIndexOf('.') + 1)
+                .equals(getMethodName());
     }
 }