+++ src/main/java/com/adobe/campaign/tests/service/testobjects/StaticType.java
@@ -1,0 +1,7 @@
+package com.adobe.campaign.tests.service.testobjects;
+
+public class StaticType {
+    public static String fetchInstantiableStringValue(Instantiable in_instantiableObject) {
+        return in_instantiableObject.getValueString();
+    }
+}